Move memstore from v2
